+++ frontend/src/pages/Contracts/Contract/ContractDetails/InvoiceSection/index.tsx
@@ -6,6 +6,13 @@ import { editFields, readFields } from "./fields";
 export const InvoiceSection = () => {
   const { contractId } = useParams();
 
+  const roles = {
+    get: "contracts_read_all",
+    add: "contracts_add_one",
+    update: "contracts_update_one",
+    delete: "contracts_delete_one",
+  };
+
   return (
     <TableData
       itemName="Invoice"
@@ -15,6 +22,7 @@ export const InvoiceSection = () => {
       createFormInitialValues={{}}
       readFields={readFields}
       editFields={editFields}
+      roles={roles}
     />
   );
 };